Rename config module
